refactor: Modal content move to own file
